+++ dune/istl/test/umfpacktest.cc
@@ -16,24 +16,6 @@
 
 using namespace Dune;
 
-// helper to add something to the diagonal of a BCRSMatrix with blocks
-auto addToDiagonal = [](auto&& matrix, const auto& value) {
-  for( auto rowIt=matrix.begin(); rowIt!=matrix.end(); rowIt++ )
-  {
-    for( auto entryIt=rowIt->begin(); entryIt!=rowIt->end(); entryIt++ )
-    {
-      if ( entryIt.index() == rowIt.index() )
-      {
-        auto&& block = Dune::Impl::asMatrix(*entryIt);
-        for (auto it=block.begin(); it!=block.end(); ++it)
-        {
-          (*it)[it.index()] += value;
-        }
-      }
-    }
-  }
-};
-
 template<typename Matrix, typename Vector, typename BitVector>
 TestSuite runUMFPack(std::size_t N)
 {
@@ -42,10 +24,8 @@ TestSuite runUMFPack(std::size_t N)
   Matrix mat;
   Vector b(N*N), x(N*N), b1(N/2), x1(N/2);
 
-  setupLaplacian(mat,N);
-
-  // make this regular
-  addToDiagonal(mat,100.0);
+  // with additional diagonal regularization
+  setupLaplacian(mat,N,100.0);
 
   b=1;
   b1=1;
@@ -148,14 +128,10 @@ TestSuite runUMFPackMultitype2x2(std::size_t N)
 
 
   // initialize all 4 matrix blocks
-  setupLaplacian(mat[_0][_0],N);
+  setupLaplacian(mat[_0][_0],N,100.0); // regularize the diagonal block
   setupLaplacian(mat[_0][_1],N);
   setupLaplacian(mat[_1][_0],N);
-  setupLaplacian(mat[_1][_1],N);
-
-  // make this regular
-  addToDiagonal(mat[_0][_0],200.0);
-  addToDiagonal(mat[_1][_1],100.0);
+  setupLaplacian(mat[_1][_1],N,200.0); // regularize the diagonal block
 
   b=1.0;
   b1=1.0;